Change the special symbols used and bump version
We feel the new symbols look more professional and also visually
pleasing. We also have a block-closing symbol now, so messages
for the combined commit can be divided into multiple blocks and
be mixed with other paragraphs in the member commit's message.

Since this is a backwards-incompatible change, we bumped the
version number to 0.2.0.